refactor: 결과지 상수 전역 및 카드별로 분리
