+++ web/src/constants/analyze/page/analyzePageConstants.ts
